unique conversion chain: T* -> unique_ptr<T> -> std::shared_ptr<T>
Squashed commit of the following:

commit 154b8c5af9f8a6c660da1c83b4a9d7930c590b7e
Author: Roman Paukner <romanpauk@gmail.com>
Date:   Sun Dec 17 20:46:42 2023 +0100

    cyclical fix

commit bef07206059552e694bca8be466810c443d914d5
Author: Roman Paukner <romanpauk@gmail.com>
Date:   Sun Dec 17 17:54:57 2023 +0100

    build fix

commit 3d47694fcee3e7394b1c87c6723e0eff4d89f8f9
Author: Roman Paukner <romanpauk@gmail.com>
Date:   Sun Dec 17 11:41:27 2023 +0100

    cleanup

commit d42576875d364e7786c1ec4ff38877d74b12e84a
Author: Roman Paukner <romanpauk@gmail.com>
Date:   Sun Dec 17 02:30:03 2023 +0100

    deallocate

commit e8ab4d10890610d69d7139f06a236217b9b44474
Author: Roman Paukner <romanpauk@gmail.com>
Date:   Sun Dec 17 01:59:44 2023 +0100

    conversions I

commit 705563491f3238e0a7a3fa6fa61604ebd4defe47
Author: Roman Paukner <romanpauk@gmail.com>
Date:   Fri Dec 15 14:54:12 2023 +0100

    additional conversions

commit a34498107b41d2c24ff1e15146133679e713d158
Author: Roman Paukner <romanpauk@gmail.com>
Date:   Thu Dec 14 23:09:01 2023 +0100

    cleanup

commit 53bfecd81b96aef00af3c46be4199bd8bc4c4601
Author: Roman Paukner <romanpauk@gmail.com>
Date:   Thu Dec 14 22:32:08 2023 +0100

    notes

commit 8291f64f07b4e767338e53956f48b3f0e95a792f
Author: Roman Paukner <romanpauk@gmail.com>
Date:   Thu Dec 14 22:20:47 2023 +0100

    tests 2

commit 770088092e6a4730e806dd7b504b53f0101090f2
Author: Roman Paukner <romanpauk@gmail.com>
Date:   Thu Dec 14 22:11:49 2023 +0100

    tests

commit df28af0a2d38a59c67cc5eafbff34e67357bc359
Author: Roman Paukner <romanpauk@gmail.com>
Date:   Thu Dec 14 11:13:02 2023 +0100

    notes

+++ test/index.cpp
@@ -58,12 +58,10 @@ template <typename Container> struct get_index_type {
 template <typename Container>
 using get_index_type_t = typename get_index_type<Container>::type;
 
-template <typename T> static T value(size_t);
-template <> int value<int>(size_t i) { return (int)i; }
-template <> size_t value<size_t>(size_t i) { return i; }
-template <> std::string value<std::string>(size_t i) {
-    return std::to_string(i);
-}
+template <typename T> static T value(int);
+template <> int value<int>(int i) { return i; }
+template <> size_t value<size_t>(int i) { return i; }
+template <> std::string value<std::string>(int i) { return std::to_string(i); }
 
 TYPED_TEST(index_test, index_tag) {
     using indexes =
